Grow ntpd/work_thread.c arrays as needed.
Add DEBUG_* variants of ntp_assert.h macros which compile away using
./configure --disable-debugging.

Fix tvalops.cpp unit test failures for 32-bit builds.Return to a
single autoreconf invocation in ./bootstrap script.Fix warnings seen
on FreeBSD 9.NMEA driver documentation update from Juergen Perlinger.

+++ util/ntp-keygen.c
@@ -271,7 +271,7 @@ main(
 	passwd2 = NULL;
 	gettimeofday(&tv, 0);
 	epoch = tv.tv_sec;
-	fstamp = epoch + JAN_1970;
+	fstamp = (u_int)(epoch + JAN_1970);
 
 	{
 		int optct = ntpOptionProcess(&ntp_keygenOptions,
